Fix exile controller (maybe?)
