+++ Submerged/ExileCutscene/Patches/BaseGameResolvingPatches.cs
@@ -6,6 +6,16 @@ namespace Submerged.ExileCutscene.Patches;
 [HarmonyPatch]
 public static class BaseGameResolvingPatches
 {
+    public static string LastExileText { get; private set; }
+
+    [HarmonyPatch(typeof(ExileController), nameof(ExileController.Begin))]
+    [HarmonyPostfix, HarmonyPriority(Priority.Last)]
+    public static void GetTextPatch(ExileController __instance)
+    {
+        LastExileText = __instance.completeString;
+        Message("Grabbing " + __instance + " from ExileController");
+    }
+
     [HarmonyPatch(typeof(ExileController), nameof(ExileController.Begin))]
     [HarmonyPrefix]
     public static void ResolvePlayerPatch(ExileController __instance)
